refactor(mcp): move reachable_url out of runtimes into mcp serving
Reachability between a service and its consumer is a cross-cutting serving
concern, not a single runtime's. After the interception unification reachable_url's
only caller is mcp serving, yet it lived in runtimes/base.py and reached back into
interception.tunnel (a lazy import to dodge a cycle). Move it next to its caller:

- relocate reachable_url to mcp/launch.py; it imports PrimeTunnel directly (no lazy
  import, no runtimes -> interception back-dependency)
- collapse the signature to reachable_url(service, port, *, colocated,
  consumer_is_local) -- drop the consumer Runtime|None duality + the
  service-is-consumer identity check; the caller passes the two bools it has.
  consumer_is_local for a tool is read off the harness runtime object (a HOST-driven
  user is local; a shared eval-level tool uses harness_is_local)
- delete the now-unused _Host / HOST sentinel (it existed only for reachable_url)
- drop the dead HOST-as-service handling + the stale "interception reachability"
  docstring; Runtime keeps only expose (the real per-runtime primitive)

Verified: ruff + 23 fast tests; /state e2e across all four reachability branches --
host colocated/own-runtime (localhost), modal tool/user/shared (expose), and
modal-harness + host-tool (tunnel) -- all pass.
