+++ verifiers/v1/runtimes/__init__.py
@@ -12,10 +12,8 @@ from typing import Annotated
 from pydantic import Field
 
 from verifiers.v1.runtimes.base import (
-    HOST,
     ProgramResult,
     Runtime,
-    reachable_url,
     register,
 )
 from verifiers.v1.runtimes.docker import DockerConfig, DockerRuntime
@@ -59,8 +57,6 @@ __all__ = [
     "RuntimeConfig",
     "make_runtime",
     "runtime_is_local",
-    "reachable_url",
-    "HOST",
     "SubprocessConfig",
     "SubprocessRuntime",
     "DockerConfig",
