+++ verifiers/v1/runtimes/base.py
@@ -270,57 +270,3 @@ class Runtime(ABC):
         `client.expose`), torn down with the sandbox in `stop()`. The reverse of a host `Tunnel`
         (which reaches a host port from inside a runtime)."""
         return None
-
-
-class _Host:
-    """The host network as a `reachable_url` location: shares the host network (so it's `is_local`)
-    and publishes nothing itself (it's reached *into* via a host `Tunnel`, not via `expose`)."""
-
-    is_local = True
-
-
-HOST = _Host()
-"""The host network, as a service location (e.g. the interception server) or a consumer (the
-framework driving a user sim) — see `reachable_url`."""
-
-
-@contextlib.asynccontextmanager
-async def reachable_url(
-    service, port: int, *, consumer=None, consumer_is_local: bool = True
-):
-    """Yield a URL for the service at (`service`, `port`) reachable from its consumer — the single
-    place tool / user / interception reachability is decided, over the two primitives `expose`
-    (publish *out* of a runtime) and a host `Tunnel` (reach *into* the host from a runtime).
-
-    `service` is the `Runtime` the service runs in, or `HOST` (a host-network service). `consumer` is
-    the consuming `Runtime` (used for the colocated check and its locality); leave it `None` for a
-    host consumer or an eval-level consumer with no single instance (a shared tool reused by every
-    rollout's harness) and pass its locality as `consumer_is_local`:
-
-    - same location (a colocated tool in the consumer's own runtime, or host -> host): localhost;
-    - the service runs in a sandbox (a remote runtime): its own published URL (`expose`), reachable
-      from anywhere;
-    - the service is on the host network: localhost to a host-network consumer, else a host tunnel
-      (a host `Tunnel`)."""
-    is_local = consumer.is_local if consumer is not None else consumer_is_local
-    if service is consumer:  # colocated in the consumer's runtime (or host -> host)
-        yield f"http://127.0.0.1:{port}"
-    elif (
-        service is not HOST and not service.is_local
-    ):  # in a sandbox → it publishes its own port
-        yield await service.expose(port)
-    elif is_local:  # host network, local consumer → localhost, no tunnel
-        yield f"http://127.0.0.1:{port}"
-    else:  # host network, remote consumer → a host tunnel publishes the port outward
-        # The prime_tunnel host bridge is a `PrimeTunnel` (interception.tunnel); imported here to
-        # avoid an import cycle (that package depends on this module). Wrap only tunnel setup as
-        # TunnelError — the exit stack scopes it, so a body error during `yield` propagates raw.
-        from verifiers.v1.errors import TunnelError
-        from verifiers.v1.interception.tunnel import PrimeTunnel
-
-        async with contextlib.AsyncExitStack() as stack:
-            try:
-                url = await stack.enter_async_context(PrimeTunnel().expose(port))
-            except Exception as e:
-                raise TunnelError(f"host tunnel (port {port}) failed: {e}") from e
-            yield url
